requirements.txt
scipy
pandas
numpy
plotly
google-cloud-storage==1.42.1
requests==2.21.0
pydub==0.25.1
webrtcvad==2.0.10
SpeechRecognition==3.8.1
librosa==0.8.1
python_speech_features==0.6
p-tqdm==1.3.3